2019 Day 9: Sensor Boost - part01

+++ 2019/src/main/java/de/adventofcode/chrisgw/day02/AdventOfCodeDay02.java
@@ -1,6 +1,6 @@
 package de.adventofcode.chrisgw.day02;
 
-import java.util.regex.Pattern;
+import de.adventofcode.chrisgw.intcode.IntCodeProgram;
 
 
 /**
@@ -15,7 +15,7 @@ public class AdventOfCodeDay02 {
     }
 
 
-    public int restoreGravityAssistProgram() {
+    public long restoreGravityAssistProgram() {
         return runIntCodeProgram(12, 2);
     }
 
@@ -23,7 +23,7 @@ public class AdventOfCodeDay02 {
     public int completeGravityAssistProgram(int whisedOutput) {
         for (int noun = 0; noun < 100; noun++) {
             for (int verb = 0; verb < 100; verb++) {
-                int output = runIntCodeProgram(noun, verb);
+                long output = runIntCodeProgram(noun, verb);
                 if (output == whisedOutput) {
                     return 100 * noun + verb;
                 }
@@ -33,7 +33,7 @@ public class AdventOfCodeDay02 {
     }
 
 
-    private int runIntCodeProgram(int noun, int verb) {
+    private long runIntCodeProgram(int noun, int verb) {
         intCodeProgram.reset();
         intCodeProgram.setNoun(noun);
         intCodeProgram.setVerb(verb);
